Switch to find-first queries for datasets
There are some datasets in DP1 that were revised, with the updated version replacing an incorrect version.  We don't want to ship the broken versions, so we now use find_first=True to pull the "best" version from the top of the collection chain.

We still use find_first=False for calibrations, since we need all of the calibration datasets for all validity ranges.